+++ ui/src/main/java/dk/digitalidentity/rc/interceptor/LdapUpdaterHook.java
@@ -139,9 +139,7 @@ public class LdapUpdaterHook implements RoleChangeHook {
 	
 	@Override
 	public void interceptAddUserRoleAssignmentOnOrgUnit(OrgUnit ou, UserRole userRole, boolean inherit) {
-		if (!orgUnitService.getUserRoles(ou, true).contains(userRole)) {
-			pendingADUpdateService.addUserRoleToQueue(userRole);
-		}
+		pendingADUpdateService.addUserRoleToQueue(userRole);
 	}
 	
 	@Override